+++ src/Exceptions/PaynlTransactionException.php
@@ -1,0 +1,13 @@
+<?php declare(strict_types=1);
+
+namespace PaynlPayment\Shopware6\Exceptions;
+
+use Exception;
+
+class PaynlTransactionException extends Exception
+{
+    public static function captureError(string $message = ''): PaynlTransactionException
+    {
+        return new PaynlTransactionException($message ?: 'Unknown PAY. transaction capture error');
+    }
+}