+++ src/PaynlPaymentShopware6.php
@@ -3,7 +3,9 @@
 namespace PaynlPayment\Shopware6;
 
 // phpcs:disable
-require_once(__DIR__ . '/../vendor/autoload.php');
+if (file_exists(__DIR__ . '/../vendor/autoload.php')) {
+    require_once(__DIR__ . '/../vendor/autoload.php');
+}
 // phpcs:enable
 
 use Doctrine\DBAL\Connection;